refactor(tests): separate tests into endpoints, clean up issues #32
